test(queue): Consolidate toadlet GET coverage
Merge the standalone QueueToadlet GET tests into QueueToadletTest and add deterministic GET-path coverage for disabled FCP, gateway denial, count-page intent parsing, placeholder replacement, and runtime-port failure mapping.

Extend LegacyQueuePagePortTest with null-request and missing-FCP edge cases for empty-page and key-list rendering.
